Add some alternative solutions to 08a
Implemented the same algorithm first in C# to confirm that it worked.
And it did, but was >10x slower than the first C++ implementation, so I
wanted to figure out if it was down to mono startup time and/or general
C# slowness so I implemented the same basic thing in C++, which runs in
roughly 75% of the time it takes the original implementation. So the
algorithm is better.
This is all very vishy washy ofc, I haven't made sure that the
difference between C#s HashSet and my long char array in C++ is what
made up the main bulk of the difference in runtime, nor the IO routines.